fix TAD Filme
