+++ include/Filme.hpp
@@ -20,41 +20,39 @@ std::map<std::string, int> Categoria = {
 };
 
 class Filme {
+    protected:
+        int identificador;
+        int unidades;
+        std::string titulo;
 
-protected:
-    int identificador;
-    int unidades;
-    std::string titulo;
-
-public:
-    Filme(int, int, std::string); 
-    virtual int calculaPreco(int) = 0; // Método virtual puro. Impede a instaciação de objetos da classe Filme diretamente
-    virtual void imprimir(); // Método para imprimir dados do DVD ou FITA, se comporta de maneira diferente em cada um
+    public:
+        Filme(int, int, std::string);
+        int getIdentificador();
+        virtual void imprimir(); // Método para imprimir dados do DVD ou FITA, se comporta de maneira diferente em cada um
 };
 
 
 // Subclasse DVD que herda todos os atributos básicos da Classe Filme
 
 class DVD : private Filme {
-private:
-    int categoria; // 0 - Lancamento, 1 - Estoque, 2 - Promocao;
-    
-public:
-    DVD(int, int, std::string, int); // Construtor
-    int calculaPreco(int) override; // Método que calcula preço do filme em função da categoria do DVD
-    void imprimir() override;
+    private:
+        int categoria; // 0 - Lancamento, 1 - Estoque, 2 - Promocao;
+        
+    public:
+        DVD(int, int, std::string, int); // Construtor
+        void imprimir() override;
 };
 
 // Subclasse FITA que herda todos os atributos básicos da Classe Filme
 
 class FITA: private Filme {
-private:
-    bool estaRebobinado;
+    private:
+        bool estaRebobinado;
 
-public:
-    FITA(int, int, std::string, bool); // Construtor
-    int calculaPreco(int) override; // Método que calcula o preço do filme levando em conta se a fita está ou não rebobinada
-    void imprimir() override;
+    public:
+        FITA(int, int, std::string, bool); // Construtor
+        bool isRebobinado();
+        void imprimir() override;
 };
 
 #endif
