+++ src/Filme.cpp
@@ -3,6 +3,8 @@
 Filme::Filme (int _unidades, int _identificador, std::string _titulo) : 
     unidades(_unidades), identificador(_identificador), titulo(_titulo) {}
 
+int Filme::getIdentificador(){ return this->identificador; }
+
 DVD::DVD (int _unidades, int _identificador, std::string _titulo, int _categoria) : 
     Filme (_unidades, _identificador, _titulo) {
         this->categoria = _categoria;
@@ -14,21 +16,7 @@ FITA::FITA (int _unidades, int _identificador, std::string _titulo, bool _estaRe
         std::cout << "Filme " << identificador << " cadastrado com sucesso" << std::endl;  
     }
 
-int DVD::calculaPreco (int dias){
-    if (categoria == LANCAMENTO) return 20 * dias;
-    
-    else if (categoria == ESTOQUE) return 10 * dias;
-
-    else if (categoria == PROMOCAO) return 10;
-
-    //Elaborar tratamento de erro futuramente;
-}
-
-// Define um valor padrão para o número de dias, pois isso não tem influência no cálculo do preço da fita;
-
-int FITA::calculaPreco (int dias = 0) { 
-    return estaRebobinado ? 5 : 7; // R$5,00 se a fita estiver rebobinada, caso contrário há uma adição de R$2,00;
-}
+bool FITA::isRebobinado(){ return this->estaRebobinado; }
 
 void Filme::imprimir() {
     std::cout << identificador << ' ' << titulo << ' ' << unidades << ' ';
